+++ src/lib/components/Form/GenericForm.svelte
@@ -62,7 +62,13 @@
 	setContext('superform', superform);
 </script>
 
-<form {method} use:enhance class={cn('flex flex-col gap-sm', className)} {...props}>
+<form
+	{method}
+	use:enhance
+	class={cn('flex w-full flex-col gap-sm', className)}
+	datatype={dataType}
+	{...props}
+>
 	{#if debug}
 		<SuperDebug data={$form}></SuperDebug>
 	{/if}