+++ src/routes/(auth)/oauth/callback/[provider]/+server.ts
@@ -3,7 +3,7 @@ import { getOAuthClient } from '$lib/server/auth/oauth/getOAuthClient';
 import { isOAuthProvider, type OAuthProvider } from '$lib/server/auth/oauth/OAuthClient';
 import { db } from '$lib/server/db';
 import { accounts, users } from '$lib/server/db/schema';
-import { createLogger } from '$lib/server/logger';
+import { createLogger } from '$lib/helpers/logger';
 import { error, redirect } from '@sveltejs/kit';
 import { eq } from 'drizzle-orm';
 import type { RequestHandler } from './$types';