+++ src/routes/(auth)/register/+page.server.ts
@@ -1,7 +1,7 @@
 import { availableOAuthProviders } from '$lib/server/auth/oauth/getOAuthClient';
 import { db } from '$lib/server/db';
 import users from '$lib/server/db/schema/user';
-import { createLogger } from '$lib/server/logger';
+
 import { hashPassword } from '$lib/server/utils';
 import { type Actions, fail, redirect } from '@sveltejs/kit';
 import { eq } from 'drizzle-orm';
@@ -10,6 +10,7 @@ import { zod } from 'sveltekit-superforms/adapters';
 import type { PageServerLoad } from './$types';
 import { StatusCodes } from 'http-status-codes';
 import { registerUserSchema } from '$lib/schemas/user';
+import { createLogger } from '$lib/helpers/logger';
 
 const logger = createLogger('register');