+++ src/routes/admin/events/(crud)/create/+page.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+	import EventsForm from '../_components/EventsForm.svelte';
+	import type { PageData } from './$types';
+
+	let { data }: { data: PageData } = $props();
+</script>
+
+<EventsForm data={data.form} />