+++ src/routes/admin/events/+layout.server.ts
@@ -1,0 +1,5 @@
+import type { LayoutServerLoad } from './$types';
+
+export const load = (async () => {
+	return { title: { text: 'Events', href: '/admin/events' } };
+}) satisfies LayoutServerLoad;