+++ src/routes/admin/submissions/+page.svelte
@@ -3,23 +3,27 @@
 	import EntityFilterBar from '../_components/EntityFilterBar.svelte';
 	import EntityList from '../_components/EntityList.svelte';
 	import type { PageData } from './$types';
+	import EventCombobox from '../../(app)/submissions/_components/EventCombobox.svelte';
 
-	export let data: PageData;
+	let { data }: { data: PageData } = $props();
 </script>
 
 <!-- TODO Figure out what to do when deleting category -->
-<EntityFilterBar entityName="submissions" />
+<EntityFilterBar entityName="submissions">
+	{#snippet buttons()}
+		<EventCombobox
+			items={data.events.map((event) => ({ label: event.name, value: event.id.toString() }))}
+		/>
+	{/snippet}
+</EntityFilterBar>
+<!-- TODO Show text filter by event -->
 <EntityList
-	items={data.submissions.flatMap((submission) => ({
-		...submission,
-		username: submission.user.username,
-		thumbnailUrl: submission.thumbnail.url,
-		category: submission.category.name
-	}))}
+	items={data.submissions}
 	fields={[
 		{ name: 'title', minScreen: 'all', sortable: true },
 		{ name: 'username', minScreen: 'lg', sortable: true },
 		{ name: 'category', minScreen: 'md', sortable: true },
+		{ name: 'event', minScreen: 'lg', sortable: true },
 		{ name: 'status', minScreen: 'sm', sortable: true }
 	]}
 	actions={[